Allow to disable all tests that require networking
Doing a test against a cloud API is not possible in some environments.
It's still useful to run a test suite in those cases, just not the
online parts of the test.

Most of the cloud APIs already require an API key, and if one is not
provided then the online part of those APIs won't be tested. The only
exception seems to be the Google APIs.

Add a SKIP_ONLINE environment variable to skip those bits.