fix: remove the requirement for defining GOOGLE_APPLICATION_CREDENTIALS
`GOOGLE_APPLICATION_CREDENTIALS` or `credential.json` is not a requirement for using Speech Recognition. `GOOGLE_APPLICATION_CREDENTIAL` is the least preferred way of authenticating with a google service. When running on Google Cloud, they inherently have the credentials attached to the infrastructure (using IAM roles), so, `GOOGLE_APPLICATION_CREDENTIALS` is not a hard requirement. At any case, if any environment variable is not defined, or `credentials.json` is not provided or detected by google cloud package, the package will produce an exception. By leaving the authentication part to `google.cloud` package, we can remove the redundant checks from Speech Recognition's code.