Add support for Faster Whisper
The Faster Whisper library purports to be significantly faster than the
original OpenAI Whisper library, so add support for it as well.

It is, unfortunately, not completely identical so we need this new
module to support it. That said 90% of it is still the same as the
original whisper, so most of that was just copied from the existing
function in __init__.py.

Note that I found in testing, that on a Pi4 with a decent SD card,
importing the "torch" library took upwards of 4 seconds, wasting
valuable time just to detect CUDA support. So instead of trying to do
this here, I added several configurable parameters from upstream to set
CUDA, etc. support if the user wants it. It will default to CPU mode
unless set.

The print statement is there for debugging to confirm the sent options,
but I don't see any sort of logger or printing inside this library.
It can be removed if desired.