Move faster_whisper transcribing to subprocess
I found during testing that faster_whisper, or at least the combination
of faster_whisper and how it is being called here, had a major memory
leak, causing the parent process to consume anywhere from 10-100MB of
memory for each run until it OOM'd. Despite multiple attempts I also
could not get the garbage collector to clean this up even by deleting
every possible object.

So to avoid this problem, I split the work into two functions. The main
recognize_faster_whisper is a stub function which simply sets up the
subprocess and a queue for interchanging data, calls and joins the
subprocess, and then returns the data passed via the queue once
completed. The bulk of the work, including the import of faster_whisper,
numpy, etc., is now performed in the subprocess function, allowing it to
be fully terminated after each run and thus truly free the memory it was
using.

In testing this works precisely as expected, with each run resulting in
zero increased memory utilization in the parent process.